refactor: some cleanup ♻️
