+++ crates/handlers/src/postgres/mod.rs
@@ -2,7 +2,7 @@ use async_trait::async_trait;
 use serde_json::Value::{self as JsonValue};
 use sqlx::{any::AnyRow, AnyPool};
 use tx_lib::handler::{Handler, RowHandler, TableHandler};
-use tx_lib::{ColumnProps, FKRows, FkRelation};
+use tx_lib::types::{ColumnProps, FKRows, FkRelation};
 
 #[derive(Debug)]
 pub struct PostgresHandler;
@@ -83,7 +83,7 @@ impl RowHandler for PostgresHandler {
         table_name: String,
         column_name: String,
         cell_value: JsonValue,
-    ) -> Result<Option<Vec<FKRows>>, String> {
+    ) -> Result<Vec<FKRows>, String> {
         let fk_relations = sqlx::query_as::<_, FkRelation>(
             "
             SELECT
@@ -107,10 +107,6 @@ impl RowHandler for PostgresHandler {
         .await
         .map_err(|err| err.to_string())?;
 
-        if fk_relations.is_empty() {
-            return Ok(None);
-        }
-
         let mut result = Vec::new();
 
         for relation in fk_relations.iter() {
@@ -132,6 +128,6 @@ impl RowHandler for PostgresHandler {
             result.push(FKRows::new(relation.table.clone(), decoded_row_data));
         }
 
-        Ok(Some(result))
+        Ok(result)
     }
 }
