+++ crates/lib/src/decode/mod.rs
@@ -5,7 +5,7 @@ use sqlx::{
     Column, Row, TypeInfo, Value, ValueRef,
 };
 
-/// Utility to decode *most* of db types into rust types.
+/// Utility to decode *most* of db types into serializable rust types.
 /// this code was taken from [here] (https://github.com/tauri-apps/tauri-plugin-sql/blob/v1/src/decode)
 pub fn to_json(v: AnyValueRef) -> Result<JsonValue, String> {
     if v.is_null() {
@@ -107,6 +107,9 @@ pub fn to_json(v: AnyValueRef) -> Result<JsonValue, String> {
     Ok(res)
 }
 
+/// Transform/Decode a `Vec<AnyRow>` into a serializable datastructure.
+///
+/// Typically used with `SELECT *`.
 pub fn decode_raw_rows(rows: Vec<AnyRow>) -> Result<Vec<JsonMap<String, JsonValue>>, String> {
     let mut result = Vec::new();
